Client

+++ .castor/client.php
@@ -34,8 +34,12 @@ function build(#[AsArgument] $toDirectory = '../client'): void
 
     rename($styleGuideFile, $styleGuideFile.'_skip');
     file_put_contents($styleGuideFile, 'export const Styleguide = () => null;');
-    run('yarn run tsc -b && yarn run vite build', context: context()->withWorkingDirectory('./client'));
+    $exitCode = run('yarn run tsc -b && yarn run vite build', context: context()->withWorkingDirectory('./client')->withAllowFailure(true))->getExitCode();
     rename($styleGuideFile.'_skip', $styleGuideFile);
+
+    if ($exitCode !== 0) {
+        throw new \RuntimeException('Build Failed.');
+    }
 
     $toDirectory = rtrim($toDirectory, '/');
     run("rm -rfv $toDirectory/*");
